Add Roadmap; add iter_reduce and iter_count
The roadmap is a document that is tracking all
of Laravel Collection's public methods. It will
get ticked off one-by-one as new functions are added.

Realistically, it will probably fall out of sync with the
code itself during development. But, the end goal is to have
each of those functions.